Align apple implementation with other native platforms: Move NativeMutexNode to a separate file and rename enter/exit to lock/unlock
